+++ domestic/helpers.py
@@ -24,26 +24,24 @@ def build_route_context(user, context={}):
 
 def get_ancestor(page, ancestor_class):
     # Seek up the tree to find a page matching class
-    while page:
-        if isinstance(page.specific, ancestor_class):
-            return page
-        page = page.get_parent()
+    return ancestor_class.objects.live().ancestor_of(page).specific().first()
 
 
 def module_has_lesson_configured_in_topic(
     module_page: CuratedListPage,
     lesson_page: DetailPage
 ) -> bool:
-    for topic_block in module_page.topics:
-        if topic_block.id == lesson_page.topic_block_id:
+    lesson_topic_page = lesson_page.get_parent().specific
+    for topic_page in module_page.get_topics():  # get_topics includes .live() by default
+        if topic_page == lesson_topic_page:
             return True
     return False
 
 
 def get_lesson_completion_status(user, context={}):
-    """Gets all lesson pages (DetailPages and uses the parental tree to get a list
-    of modules (CuratedListPages), with some filtering-out of DetailPages which
-    are not associated with a CuratedListPage.topics field
+    """Gets all lesson pages (DetailPages) and uses the parental tree to get a list
+    of modules (CuratedListPages), with some grouping of DetailPages by their parent
+    TopicPage (which itself is a child of CuratedListPage)
 
     Example output:
     {
@@ -52,7 +50,7 @@ def get_lesson_completion_status(user, context={}):
             {
                 'total_pages': 7,
                 'completion_count': 4,
-                'page': <Page: Identify opportunities and research the market>,
+                'page': <CuratedListPage: Identify opportunities>,
                 'completed_lesson_pages': defaultdict(
                     <class 'set'>, {
                     'b7eca1bf-8b43-4737-91e4-913dfeb2c5d8': {10, 26},
@@ -76,15 +74,8 @@ def get_lesson_completion_status(user, context={}):
 
     page_map = {}
 
-    for detail_page in DetailPage.objects.live().filter(
-        topic_block_id__isnull=False
-        # ie: get ONLY detail pages that are mapped to a topic
-        # (This mapping happens via a wagtail_hook, which sets topic_block_id
-        # on the appropriate DetailPage for each topic in CuratedListPage.topics)
-
-        # This ALSO means that CuratedListPages with no topics set up, even if
-        # they have child DetailPages, will NOT be included in these results.
-    ):
+    # TODO: refactor further to simplify
+    for detail_page in DetailPage.objects.live():
         module_page = get_ancestor(detail_page, CuratedListPage)
         if module_page:
             page_map[module_page.id] = (
@@ -96,7 +87,8 @@ def get_lesson_completion_status(user, context={}):
                 }
             )
 
-            # Only proceeed with `detail_page` if it is CURRENTLY configured in a topic
+            # Only proceeed with `detail_page` if it is CURRENTLY
+            # configured in a topic for this module
             if not module_has_lesson_configured_in_topic(
                 module_page=module_page.specific,
                 lesson_page=detail_page.specific
@@ -106,7 +98,7 @@ def get_lesson_completion_status(user, context={}):
             page_map[module_page.id]['total_pages'] += 1
 
             if detail_page.id in completed:
-                topic_id_as_key = detail_page.topic_block_id
+                topic_id_as_key = detail_page.get_parent().id
                 page_map[module_page.id]['completed_lesson_pages'][topic_id_as_key].add(
                     detail_page.id
                 )